Optimise the use of an attribute set.
Whenever the attribute set objects were used, we allocated excessive
memory, occupied larger registers, performed wider operations with
those when we didn't need to do that and avoid the extra costs of those.
At the same time, the postfix increments were used instead of the prefix
ones.

This commit brings those to justice and attempts to avoid the overhead,
providing the compiler with more precise information about the types
used and doing only what is required in the code.

+++ src/graph/entities/attribute_set.h
@@ -25,7 +25,7 @@ typedef struct {
 } Attribute;
 
 typedef struct {
-	uint16_t attr_count;       // number of attributes
+	uint16_t attr_count;     // number of attributes
 	Attribute attributes[];  // key value pair of attributes
 } _AttributeSet;
 
@@ -44,7 +44,7 @@ SIValue *AttributeSet_Get
 SIValue AttributeSet_GetIdx
 (
 	const AttributeSet set,  // set to retieve attribute from
-	int i,                   // index of the property
+	uint16_t i,              // index of the property
 	Attribute_ID *attr_id    // attribute identifier
 );
